requirements.txt
autogen-agentchat>=0.7.5
autogen-ext[openai]>=0.7.5
python-dotenv>=1.0.0
PyMuPDF>=1.23.0